pyproject.toml
[project]
name = "ai-devops-agent"
version = "0.1.0"
description = "AI DevOps Agent MVP — self-contained demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "pydantic>=2.9",
    "openai>=1.54",
    "python-dotenv>=1.0",
    "sse-starlette>=2.1",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*", "sample_service*"]
